pyproject.toml
[project]
name = "askdoc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "celery>=5.6.3",
    "faiss-cpu>=1.13.2",
    "fastapi[standard]>=0.136.0",
    "openai>=2.32.0",
    "pgvector>=0.4.2",
    "psycopg2-binary>=2.9.12",
    "pydantic-settings>=2.14.0",
    "pypdf>=6.10.2",
    "python-docx>=1.2.0",
    "python-multipart>=0.0.26",
    "redis>=7.4.0",
    "sentence-transformers>=5.4.1",
    "sqlalchemy>=2.0.49",
    "langchain-text-splitters>=1.1.2",
]

[tool.uv.sources]
torch = { index = "pytorch-cpu" }

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true
